refactor: ts and misc other stuff
