+++ src/cli/fruster-config.ts
@@ -1,14 +1,13 @@
 #!/usr/bin/env node
-const { program } = require("commander");
+import { program } from "commander";
 
 program
-	.command("apply", "apply config from service registry")
 	.command("get", "get config of a service")
 	.command("set", "set config for a service")
 	.command("unset", "remove config for a service")
 	.description(
 		`
-Manage kubernetes service config.
+Manage config for apps.
 `
 	)
 	.parse(process.argv);
