+++ src/cli/fruster-destroy.ts
@@ -1,14 +1,14 @@
 #!/usr/bin/env node
 
-const { program } = require("commander");
-const { deleteDeployment, deleteSecret, deleteService, deleteReplicaSet } = require("../kube/kube-client");
-const log = require("../log");
-const { validateRequiredArg, confirmPrompt } = require("../utils/cli-utils");
+import { program } from "commander";
+import { deleteDeployment, deleteReplicaSet, deleteService } from "../kube/kube-client";
+import * as log from "../log";
+import { confirmPrompt, validateRequiredArg } from "../utils/cli-utils";
 
 program
 	.description(
 		`
-Removes an app/service and all resources related to it.
+Removes an app and all resources related to it.
 
 Examples:
 
@@ -32,8 +32,6 @@ async function run() {
 		if (confirm) {
 			await deleteDeployment(namespace, app);
 			log.info("Removed deployment");
-			await deleteSecret(namespace, app + "-config");
-			log.info("Removed config");
 			await deleteReplicaSet(namespace, app);
 			log.info("Removed replica set");
 			if (await deleteService(namespace, app)) {
