+++ src/cli/fruster-scale.ts
@@ -1,8 +1,8 @@
 #!/usr/bin/env node
 import { program } from "commander";
-import { scaleDeployment, getPods } from "../kube/kube-client";
+import { getPods, scaleDeployment } from "../kube/kube-client";
 import * as log from "../log";
-const { validateRequiredArg, getOrSelectNamespaceForApp } = require("../utils/cli-utils");
+import { validateRequiredArg } from "../utils/cli-utils";
 
 program
 	.description(
@@ -16,7 +16,7 @@ $ fctl scale -r 2 -a api-gateway
 	)
 	.option("-r, --replicas <replicas>", "number of replicas")
 	.option("-n, --namespace <namespace>", "kubernetes namespace service is in")
-	.option("-a, --app <serviceName>", "name of service")
+	.option("-a, --app <app>", "name of service")
 	.parse(process.argv);
 
 const replicas = program.getOptionValue("replicas");
@@ -24,13 +24,10 @@ const serviceName = program.getOptionValue("app");
 let namespace = program.getOptionValue("namespace");
 
 validateRequiredArg(serviceName, program, "Missing app name");
+validateRequiredArg(namespace, program, "Missing namespace");
 validateRequiredArg(replicas, program, "Missing number of replicas");
 
 async function run() {
-	if (!namespace) {
-		namespace = await getOrSelectNamespaceForApp(serviceName);
-	}
-
 	if (replicas === undefined) {
 		const pods = await getPods(namespace, serviceName);
 		log.info(`Service has ${pods.length} pods, set --replicas to scale replicas`);
